requirements.txt
numpy==1.19.3
tensorflow-cpu==2.5.0
streamlit==0.80.0
Pillow==8.3.1
